cursor/pyproject.toml
[tool.poetry]
name = "trade-latency-monitor"
version = "1.0.0"
description = "Hyperliquid Trade Latency Monitor Script"
authors = ["Your Name <your.email@example.com>"]
readme = "README_SCRIPTS.md"
packages = [{include = "trade_latency_monitor"}]

[tool.poetry.dependencies]
python = "^3.9.0"
hyperliquid-python-sdk = "^0.18.0"
watchdog = "^2.1.0"
eth-account = "^0.13.7"
eth-keys = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[tool.poetry.group.optional.dependencies]
# Chart functionality
matplotlib = "^3.5.0"
pandas = "^1.3.0"

# Alert functionality
requests = "^2.25.0"

# Performance optimization (Linux)
pyinotify = {version = "^0.9.6", platform = "linux"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
trade-monitor = "trade_latency_monitor:main"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
